+++ CHANGELOG.md
@@ -93,6 +93,7 @@
 ## RELEASE 1.5.18 - 2017-03-29
 ### Added
 - Smart Actions - Users don't have to select records to use a smart action through the global option.
+- Alerts - Support checks on record creation of a specific collection.
 
 ### Fixed
 - Analyzer - Avoid class name conflict with the "Application" model name.